Other columns are now taken into consideration
Before this update, ALL rows where modified while moving a single row without any condition. Therefore, if there was several lists inside a table, it was updating all rows instead of updating only the ones that are on the same list.

**For example**:
- `post_images` table:
  - there is a column called `post_id` referencing `posts` table.
  - when ordering `post_id=1` images, ALL images where modified instead of only images belonging to `post_id=1`

The downside is that now you have to pass the conditions to some methods.